[refactor] messageInfo 공통 타입 분리
