+++ src/app/api/chat/messages/types.ts
@@ -1,10 +1,4 @@
 import { CommonResponse } from "../../types";
+import { MessageInfo } from "../types";
 
 export type GetMessageResponse = CommonResponse<MessageInfo[]>;
-
-export interface MessageInfo {
-  message: string;
-  readBy: string[];
-  senderId: string;
-  createdAt: string;
-}
