+++ src/app/api/chat/types.ts
@@ -1,0 +1,8 @@
+import { Timestamp } from "@/types/chat";
+
+export interface MessageInfo {
+  senderId: string;
+  message: string;
+  createdAt: Timestamp;
+  readBy: string[];
+}
